refactor(daemon-local): decouple contract types from sidecar-lifecycle feature
Pure enum / error contract types (DaemonBootstrapError, ProbeOutcome,
SpawnReason, TerminateDaemonError, terminate_local_daemon_pid) lived
in daemon_bootstrap.rs / daemon_lifecycle.rs, gated behind the
sidecar-lifecycle feature only because they shared a module with
CommandChild-bearing code. Enabling sidecar-lifecycle pulls in
tauri-plugin-shell, so non-Tauri consumers (uc-desktop, future
native shells) couldn't use these types without violating the
GUI-framework agnostic boundary.

Move them into two new feature-gate-free modules:
- src/contract.rs — DaemonBootstrapError, ProbeOutcome, SpawnReason,
  TerminateDaemonError, terminate_local_daemon_pid
- src/health_wait.rs — wait_for_daemon_health, wait_for_endpoint_absent
  (probe-only, no CommandChild dependency)

Promote thiserror and tokio from optional to default dependencies —
both are pure utility crates with no GUI footprint.

Existing import paths (uc_daemon_local::daemon_bootstrap::*,
uc_daemon_local::daemon_lifecycle::*) still resolve via re-exports;
new code should target uc_daemon_local::contract::* directly.

Verified cargo tree -p uc-daemon-local -e normal shows zero
tauri-plugin edges on the default feature set.
